backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.9
django-cors-headers==4.7.0
xhtml2pdf==0.2.17
Pillow==11.2.1
